Add EntityDyeEvent. Adds BUKKIT-3900
Previously only sheep were handled for being dyed. However, currently
there are other mobs that can be dyed by the player, such as wolves. This
commit adds the event required for detecting those changes as well as
manipulates the existing event to be deprecated and implement this new
event.
