+++ src/main/java/org/bukkit/event/entity/SheepDyeWoolEvent.java
@@ -2,52 +2,24 @@ package org.bukkit.event.entity;
 
 import org.bukkit.DyeColor;
 import org.bukkit.entity.Sheep;
-import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
 
 /**
  * Called when a sheep's wool is dyed
+ *
+ * @deprecated Use EntityDyeEvent instead
  */
-public class SheepDyeWoolEvent extends EntityEvent implements Cancellable {
+@Deprecated
+public class SheepDyeWoolEvent extends EntityDyeEvent {
     private static final HandlerList handlers = new HandlerList();
-    private boolean cancel;
-    private DyeColor color;
 
     public SheepDyeWoolEvent(final Sheep sheep, final DyeColor color) {
-        super(sheep);
-        this.cancel = false;
-        this.color = color;
-    }
-
-    public boolean isCancelled() {
-        return cancel;
-    }
-
-    public void setCancelled(boolean cancel) {
-        this.cancel = cancel;
+        super(sheep, color);
     }
 
     @Override
     public Sheep getEntity() {
         return (Sheep) entity;
-    }
-
-    /**
-     * Gets the DyeColor the sheep is being dyed
-     *
-     * @return the DyeColor the sheep is being dyed
-     */
-    public DyeColor getColor() {
-        return color;
-    }
-
-    /**
-     * Sets the DyeColor the sheep is being dyed
-     *
-     * @param color the DyeColor the sheep will be dyed
-     */
-    public void setColor(DyeColor color) {
-        this.color = color;
     }
 
     @Override
@@ -58,5 +30,4 @@ public class SheepDyeWoolEvent extends EntityEvent implements Cancellable {
     public static HandlerList getHandlerList() {
         return handlers;
     }
-
 }
